+++ src/Plainion.GraphViz.Modules.MdFiles/Dependencies/Analyzer/Parser/IMarkdownParser.cs
@@ -1,6 +1,6 @@
-﻿using Plainion.GraphViz.Modules.MdFiles.Dependencies.Markdown;
+﻿using Plainion.GraphViz.Modules.MdFiles.Dependencies.Analyzer.Markdown;
 
-namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Parser
+namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Analyzer.Parser
 {
     internal interface IMarkdownParser
     {
