+++ src/Plainion.GraphViz.Modules.MdFiles/Dependencies/Analyzer/Resolver/InternalLink.cs
@@ -1,4 +1,4 @@
-﻿namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Resolver
+﻿namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Analyzer.Resolver
 {
     /// <summary>
     /// Contains the absolute path to an internal link.
