+++ src/Plainion.GraphViz.Modules.MdFiles/Dependencies/Analyzer/Resolver/LinkResolver.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 
-namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Resolver
+namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Analyzer.Resolver
 {
     internal class LinkResolver : ILinkResolver
     {
