+++ src/Plainion.GraphViz.Modules.MdFiles/Dependencies/Analyzer/Verifier/VerifiedLink.cs
@@ -1,4 +1,4 @@
-﻿namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Verifier
+﻿namespace Plainion.GraphViz.Modules.MdFiles.Dependencies.Analyzer.Verifier
 {
     internal abstract class VerifiedLink
     {
